FEAT: Implement `--stop-on-*` cli options
These options allow the user to stop the execution of multiple tests
when one of the fails (`--stop-on-failure`), throws an error
(`--stop-on-error`) or does either of the previous cases
(`--stop-on-defect`).

The included end to end tests for the async mode are still in draft mode
as the success detection via regular expressions or format string is not
flexible enough.